+++ backend/pokemon/src/main/java/net/pokeretro/pokemon/repository/PokemonRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.pokeretro.pokemon.entity.Pokemon;
 
-public interface PokemonRepository extends JpaRepository<Pokemon, Long> {
+public interface PokemonRepository extends JpaRepository<Pokemon, Integer> {
 }
